Move encoding algorithm into enigma class
