--- lib/encoding_algorithm.rb
@@ -1,7 +1,0 @@
-class EncodingAlgorithm
-attr_reader :characters
-
-def initialize(characters = ("a".."z").to_a << " ")
-  @characters = characters
-end
-end
