--- test/encoding_algorithm_test.rb
@@ -1,18 +1,0 @@
-require "./test/test_helper"
-
-class EncodingAlgorithmTest < Minitest::Test
-  def setup
-    @encoding_algorithm = EncodingAlgorithm.new
-  end
-
-  def test_it_exists
-    assert_instance_of EncodingAlgorithm, @encoding_algorithm
-  end
-
-  def test_it_has_readable_attributes
-    expected = ["a", "b", "c", "d", "e", "f", "g", "h", "i",
-                "j", "k", "l", "m", "n", "o", "p", "q", "r",
-                "s", "t", "u", "v", "w", "x", "y", "z", " "]
-    assert_equal expected, @encoding_algorithm.characters
-  end
-end
